Refactor PersonCard and PersonListPanel
